requirements.txt
Flask
gunicorn
requests
flask-bootstrap
Flask-WTF
Pillow
tensorflow>=2.18.0
keras
lxml
matplotlib
numpy
